refactor(core): decompose lowlevel module
[no changelog]
