+++ core/embed/models/T2B1/boards/trezor_r_v3.h
@@ -5,6 +5,7 @@
 
 #define USE_BUTTON 1
 #define USE_SBU 1
+#define USE_PVD 1
 
 #define DISPLAY_RESX 128
 #define DISPLAY_RESY 128
