+++ core/embed/trezorhal/stm32u5/pvd.c
@@ -1,0 +1,1 @@
+../stm32f4/pvd.c
