+++ core/site_scons/models/stm32f4_common.py
@@ -48,10 +48,12 @@ def stm32f4_common_files(env, defines, sources, paths):
         "embed/trezorhal/stm32f4/flash_otp.c",
         "embed/trezorhal/stm32f4/fwutils.c",
         "embed/trezorhal/stm32f4/layout.c",
-        "embed/trezorhal/stm32f4/lowlevel.c",
         "embed/trezorhal/stm32f4/monoctr.c",
         "embed/trezorhal/stm32f4/mpu.c",
+        "embed/trezorhal/stm32f4/option_bytes.c",
+        "embed/trezorhal/stm32f4/pvd.c",
         "embed/trezorhal/stm32f4/random_delays.c",
+        "embed/trezorhal/stm32f4/reset_flags.c",
         "embed/trezorhal/stm32f4/rng.c",
         "embed/trezorhal/stm32f4/secret.c",
         "embed/trezorhal/stm32f4/startup_init.c",
