+++ core/site_scons/models/stm32u5_common.py
@@ -57,11 +57,13 @@ def stm32u5_common_files(env, defines, sources, paths):
         "embed/trezorhal/stm32u5/flash_otp.c",
         "embed/trezorhal/stm32u5/fwutils.c",
         "embed/trezorhal/stm32u5/layout.c",
-        "embed/trezorhal/stm32u5/lowlevel.c",
         "embed/trezorhal/stm32u5/hash_processor.c",
         "embed/trezorhal/stm32u5/monoctr.c",
         "embed/trezorhal/stm32u5/mpu.c",
+        "embed/trezorhal/stm32u5/option_bytes.c",
+        "embed/trezorhal/stm32u5/pvd.c",
         "embed/trezorhal/stm32u5/random_delays.c",
+        "embed/trezorhal/stm32u5/reset_flags.c",
         "embed/trezorhal/stm32u5/rng.c",
         "embed/trezorhal/stm32u5/secret.c",
         "embed/trezorhal/stm32u5/secure_aes.c",
